refactor sitetree side navigation util
