+++ accelerator/sitetree_navigation/__init__.py
@@ -1,0 +1,7 @@
+from accelerator.sitetree_navigation.side_nav_definitions import (
+    SIDE_NAV_ITEM_PROPS_LIST,
+)
+from accelerator.sitetree_navigation.utils import (
+  create_items,
+  add_user_roles_to_side_nav_items,
+)
